requirements.txt
Django==2.0
django-appconf==1.0.2
django-cors-headers==2.1.0
django-imagekit==4.0.2
django-rest-framework==0.1.0
djangorestframework==3.7.3
imagekit==1.0.6
olefile==0.44
pilkit==2.0
Pillow==4.3.0
pytz==2017.3
six==1.11.0
